refactor step "I have multiple income and expense categories"
